// examples/openapi-ts-router/hono/petstore/src/gen/v1.ts
export interface components {
	schemas: {
		Order: {
			/**
			 * Format: int64
			 * @example 10
			 */
			id?: number;
			/**
			 * Format: int64
			 * @example 198772
			 */
			petId?: number;
			/**
			 * Format: int32
			 * @example 7
			 */
			quantity?: number;
			/** Format: date-time */
			shipDate?: string;
			/**
			 * @description Order Status
			 * @example approved
			 * @enum {string}
			 */
			status?: 'placed' | 'approved' | 'delivered';
			complete?: boolean;
		};
		Customer: {
			/**
			 * Format: int64
			 * @example 100000
			 */
			id?: number;
			/** @example fehguy */
			username?: string;
			address?: components['schemas']['Address'][];
		};
		Address: {
			/** @example 437 Lytton */
			street?: string;
			/** @example Palo Alto */
			city?: string;
			/** @example CA */
			state?: string;
			/** @example 94301 */
			zip?: string;
		};
		Category: {
			/**
			 * Format: int64
			 * @example 1
			 */
			id?: number;
			/** @example Dogs */
			name?: string;
		};
		User: {
			/**
			 * Format: int64
			 * @example 10
			 */
			id?: number;
			/** @example theUser */
			username?: string;
			/** @example John */
			firstName?: string;
			/** @example James */
			lastName?: string;
			/** @example john@email.com */
			email?: string;
			/** @example 12345 */
			password?: string;
			/** @example 12345 */
			phone?: string;
			/**
			 * Format: int32
			 * @description User Status
			 * @example 1
			 */
			userStatus?: number;
		};
		Tag: {
			/** Format: int64 */
			id?: number;
			name?: string;
		};
		Pet: {
			/**
			 * Format: int64
			 * @example 10
			 */
			id?: number;
			/** @example doggie */
			name: string;
			category?: components['schemas']['Category'];
			photoUrls: string[];
			tags?: components['schemas']['Tag'][];
			/**
			 * @description pet status in the store
			 * @enum {string}
			 */
			status?: 'available' | 'pending' | 'sold';
		};
		ApiResponse: {
			/** Format: int32 */
			code?: number;
			type?: string;
			message?: string;
		};
		/**
		 * @description Error Response
		 * @example {
		 *       "error_code": "400",
		 *       "error_description": "Bad Request",
		 *       "error_uri": null,
		 *       "additional_errors": []
		 *     }
		 */
		AppError: {
			/** @description Error code */
			error_code?: string;
			/** @description Error description */
			error_description?: string | null;
			/** @description Error URI */
			error_uri?: string | null;
			additional_errors?: Record<string, never>[];
		};
	};
	responses: never;
	parameters: never;
	requestBodies: {
		/** @description Pet object that needs to be added to the store */
		Pet: {
			content: {
				'application/json': components['schemas']['Pet'];
				'application/xml': components['schemas']['Pet'];
			};
		};
		/** @description List of user object */
		UserArray: {
			content: {
				'application/json': components['schemas']['User'][];
			};
		};
	};
	headers: never;
	pathItems: never;
}